requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary>=2.8
beautifulsoup4
djangorestframework
requests
selenium==4.1.3